DSTEW-147: Decouple openAPI development from backend API code (#6)
Added a new workflow to allow changes to the OpenAPI spec to be made and shared with clients/stakeholders without needing the backend API code to support them.

Every time a push is made to a branch named like openapi/*, a stub deployment will be created/updated in UAT with just the swagger-UI documentation updated. This can then be shared with clients and fine-tuned independently of the backend API and merged into another branch only when ready to be implemented.